feat(datatype): Add support for datatype.
BREAKING CHANGE: Support for datatypes changed the structure of the resulting struct, having default
'type' and 'def' fields. The former defines the name of what was parsed (the T_xxx, P_xxx, etc),
while the latter has custom configurations according to the context. Tests are changed to support
this new structure.