refactor: Separate project structure according to SQL dialect folder.
A dialect can now be given to the Parser constructor. The parser will build the grammar rules
according to the given dialect. Other changes had to be made such as in the npm scripts. Example
script was removed. Lib/compiled folder was renamed to Lib/mysql. Compiled grammar files are now in
lib/mysql/parser.