feat: Add timestamp initialization.
BREAKING CHANGE: Foreign key reference had to be moved out of the O_COLUMN_DEFINITION rule, moving
it out of column options. This had to be done because of grammar ambiguity when parsing "ON DELETE"
or "ON UPDATE" clauses, as it conflicted with the foreign key triggers. Test files were also updated
to ease on the job of updating expected JSON results.