fix: timestamp initialization with triggers
BREAKING CHANGE: It is a breaking change only for those who used the parsed format directly. It will not affect the compact JSON format nor the JSON Schema format. More details: foreign key reference had to be moved out of the O_COLUMN_DEFINITION rule, moving it out of column options. This had to be done because of grammar ambiguity when parsing "ON DELETE" or "ON UPDATE" clauses, as it conflicted with the foreign key triggers. It is still compliant with MySQL and MariaDB syntaxes.