feat: Add option to not use $ref in JSON Schema output.
BREAKING CHANGE: The `toJsonSchemaArray` method used to have only one argument: the tables array in
compact JSON format. Now this array is the second argument to be passed to this method. The first
argument refers to JSON Schema output options. Refer to the README for the list of options.

fix #36